+++ lib/exampleGenerator.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Hoek = require('hoek');
-const Joi = require('./Joi');
+const Joi = require('./joi');
 const ValueGenerator = require('./helpers').valueGenerator;
 
 const exampleGenerator = function (schema, options) {
