+++ lib/helpers.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Hoek = require('hoek');
-const Joi  = require('./Joi');
+const Joi  = require('./joi');
 const RandExp = require('randexp');
 const Uuid = require('uuid');
 const Alloc = require('buffer-alloc');
