+++ lib/joi.js
