+++ lib/joiGenerator.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Hoek = require('hoek');
-const Joi  = require('./Joi');
+const Joi  = require('./joi');
 
 const internals = {};
